deep_value_asset_stocks/requirements.txt
opendartreader
pykrx
pandas
requests
python-dotenv
openai
google-generativeai
pyTelegramBotAPI
